roles/mrvpn2/files/ipt-server/requirements.txt
websockets
pyroute2
pydantic
pydantic-settings
PyYaml
jinja2
